⚗️testing workflow

+++ test/TestAdoption.sol
@@ -6,23 +6,27 @@ import {AnimalFeeding} from "../contracts/AnimalFeeding.sol";
 
 contract TestAdoption {
     // defined in Adoption contract
-    uint constant MAX_ANIMAL= 10; // max number of available animals.
+    uint256 constant MAX_ANIMAL = 10; // max number of available animals.
     // Instantiate global contract
-    AnimalFeeding adoption= AnimalFeeding(DeployedAddresses.AnimalFeeding());
+    AnimalFeeding adoption = AnimalFeeding(DeployedAddresses.AnimalFeeding());
 
     function testAdoptAnimal() public {
         // Before adoption, adopters is full fo null address
-        address[] memory expectedAdopters= new address[](MAX_ANIMAL);
+        address[] memory expectedAdopters = new address[](MAX_ANIMAL);
 
-        for (uint i= 0; i < MAX_ANIMAL; i++) {
-            Assert.equal(address(0), adoption.adopters(i),
+        for (uint256 i = 0; i < MAX_ANIMAL; i++) {
+            Assert.equal(
+                address(0),
+                adoption.adopters(i),
                 "Should only contain null addresses."
             );
         }
         // After adoption, there should be one adopter address.
         adoption.adoptAnimal(0, "Alice");
-        expectedAdopters[0]= address(this);
-        Assert.equal(address(this), adoption.adopters(0),
+        expectedAdopters[0] = address(this);
+        Assert.equal(
+            address(this),
+            adoption.adopters(0),
             "First pet should have an owner."
         );
     }
@@ -32,18 +36,21 @@ contract TestAdoption {
         // raise an error if called after adoption.adoptAnimal(1).
         adoption.adoptAnimal(1, "Bob");
     }
+
     function testAdoptAnimalAlreadyAdopted() public {
         bool r;
         // Once called should work
         adoption.adoptAnimal(1, "Bob");
         // Again for same _petId=1 should not work
-        (r, ) = address(this).call(abi.encodePacked(this.raiseAdoptAnimalAlreadyAdopted.selector));
+        (r, ) = address(this).call(
+            abi.encodePacked(this.raiseAdoptAnimalAlreadyAdopted.selector)
+        );
         Assert.isFalse(r, "Should not adopt same _petId twice.");
     }
 
     function testGetAvailablePetId() public {
-        uint[] memory resultAvailablePetId;
-        uint length;
+        uint256[] memory resultAvailablePetId;
+        uint256 length;
         adoption.adoptAnimal(2, "Alice");
         adoption.adoptAnimal(3, "Alice");
         adoption.adoptAnimal(4, "Alice");
@@ -51,7 +58,7 @@ contract TestAdoption {
         adoption.adoptAnimal(6, "Alice");
         adoption.adoptAnimal(7, "Alice");
         adoption.adoptAnimal(8, "Alice");
-        (resultAvailablePetId, length)= adoption.getAvailablePetId();
+        (resultAvailablePetId, length) = adoption.getAvailablePetId();
         Assert.equal(length, 1, "Should stay only one pet.");
         Assert.equal(resultAvailablePetId[0], 9, "Should stay the 10-th pet.");
     }
