+++ test/TestAnimalFeeding.sol
@@ -4,18 +4,23 @@ import {Assert} from "truffle/Assert.sol";
 import {DeployedAddresses} from "truffle/DeployedAddresses.sol";
 import {AnimalFeeding} from "../contracts/AnimalFeeding.sol";
 
-
 contract TestAnimalFeeding {
-    uint constant MAX_ANIMAL= 10; // max number of available animals.
-    AnimalFeeding animalFeeding= AnimalFeeding(DeployedAddresses.AnimalFeeding());
+    uint256 constant MAX_ANIMAL = 10; // max number of available animals.
+    AnimalFeeding animalFeeding =
+        AnimalFeeding(DeployedAddresses.AnimalFeeding());
 
     function _setCoolDown() public {
         animalFeeding.setCoolDown(1 days);
     }
 
     function testSetCoolDownNotWork() public {
-        (bool r, ) = address(this).call(abi.encodePacked(this._setCoolDown.selector));
-        Assert.isFalse(r, "Should not change cooldown state variable if not called by owner.");
+        (bool r, ) = address(this).call(
+            abi.encodePacked(this._setCoolDown.selector)
+        );
+        Assert.isFalse(
+            r,
+            "Should not change cooldown state variable if not called by owner."
+        );
     }
 
     function _setFoodCost() public {
@@ -23,14 +28,19 @@ contract TestAnimalFeeding {
     }
 
     function testSetFoodCostNotWork() public {
-        (bool r, ) = address(this).call(abi.encodePacked(this._setFoodCost.selector));
-        Assert.isFalse(r, "Should not change food cost state variable if not called by owner.");
+        (bool r, ) = address(this).call(
+            abi.encodePacked(this._setFoodCost.selector)
+        );
+        Assert.isFalse(
+            r,
+            "Should not change food cost state variable if not called by owner."
+        );
     }
 
     function testPetAlive() public {
         bool alive;
         animalFeeding.adoptAnimal(0, "Alice");
-        (,,,, alive)= animalFeeding.animals(0);
+        (, , , , alive) = animalFeeding.animals(0);
         require(alive, "The new pet should be alive.");
     }
 
@@ -40,7 +50,7 @@ contract TestAnimalFeeding {
         bool alive;
 
         animalFeeding.adoptAnimal(1, "Alice");
-        (,,, lastFeed,)= animalFeeding.animals(1);
+        (, , , lastFeed, ) = animalFeeding.animals(1);
         //lastFeed= uint32(block.timestamp - 30 days);
         //alive= animalFeeding.isPetAlive(1);
         //require(!alive, "The new pet should not be alive.");
@@ -49,12 +59,10 @@ contract TestAnimalFeeding {
     // Not tested yet, can't mock block.timestamp
     function testFeedAnimalAlive() public {
         // We already have adopted 2 animals {0, 1}
-
     }
 
     // Not tested yet, can't mock block.timestamp
     function testFeedAnimalDead() public {
         // We already have adopted 2 animals {0, 1}
-
     }
 }
